use existing test helper function
