+++ crates/net/network/src/test_utils/mod.rs
@@ -2,9 +2,11 @@
 
 mod init;
 mod testnet;
+pub mod transactions;
 
 pub use init::{
     enr_to_peer_id, unused_port, unused_tcp_addr, unused_tcp_and_udp_port, unused_tcp_udp,
     unused_udp_addr, unused_udp_port, GETH_TIMEOUT,
 };
 pub use testnet::{NetworkEventStream, Peer, PeerConfig, PeerHandle, Testnet, TestnetHandle};
+pub use transactions::{buffer_hash_to_tx_fetcher, new_mock_session, new_tx_manager};
